LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGLength-ch-expected.txt


PASS ch unit in SVGLength
PASS Convert back to ch from new user unit value
FAIL upright vertical ch unit in SVGLength assert_approx_equals: expected 200 +/- 0.1 but got 100
